+++ app/src/main/java/by/klnvch/link5dots/domain/models/GameScore.kt
@@ -24,9 +24,9 @@
 
 package by.klnvch.link5dots.domain.models
 
-open class SimpleGameScore(val size: Int, val duration: Long, val timestamp: Long)
+open class SimpleGameScore(val size: Int, val duration: Int, val timestamp: Long)
 
-class BotGameScore(size: Int, duration: Long, timestamp: Long, val status: GameResult) :
+class BotGameScore(size: Int, duration: Int, timestamp: Long, val status: GameResult) :
     SimpleGameScore(size, duration, timestamp)
 
 enum class GameResult { WON, LOST }
